+++ framework/Furion.Extras.Authentication.JwtBearer/Extensions/JWTAuthorizationServiceCollectionExtensions.cs
@@ -2,12 +2,15 @@
 using Furion.DataEncryption;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System;
+using System.Linq;
+using System.Reflection;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -21,20 +24,26 @@ namespace Microsoft.Extensions.DependencyInjection
         /// </summary>
         /// <param name="authenticationBuilder"></param>
         /// <param name="tokenValidationParameters">token 验证参数</param>
+        /// <param name="jwtBearerConfigure"></param>
         /// <param name="enableGlobalAuthorize">启动全局授权</param>
         /// <returns></returns>
-        public static AuthenticationBuilder AddJwt(this AuthenticationBuilder authenticationBuilder, object tokenValidationParameters = default, bool enableGlobalAuthorize = false)
+        public static AuthenticationBuilder AddJwt(this AuthenticationBuilder authenticationBuilder, object tokenValidationParameters = default, Action<JwtBearerOptions> jwtBearerConfigure = null, bool enableGlobalAuthorize = false)
         {
             var services = authenticationBuilder.Services;
 
             // 配置 JWT 选项
             ConfigureJWTOptions(services);
 
+            // 获取配置选项
             var jwtSettings = services.BuildServiceProvider().GetService<IOptions<JWTSettingsOptions>>().Value;
 
+            // 添加授权
             authenticationBuilder.AddJwtBearer(options =>
             {
                 options.TokenValidationParameters = (tokenValidationParameters as TokenValidationParameters) ?? JWTEncryption.CreateTokenValidationParameters(jwtSettings);
+
+                // 添加自定义配置
+                jwtBearerConfigure?.Invoke(options);
             });
 
             //启用全局授权
@@ -53,10 +62,11 @@ namespace Microsoft.Extensions.DependencyInjection
         /// 添加 JWT 授权
         /// </summary>
         /// <param name="services"></param>
-        /// <param name="configureOptions">授权配置</param>
+        /// <param name="authenticationConfigure">授权配置</param>
         /// <param name="tokenValidationParameters">token 验证参数</param>
+        /// <param name="jwtBearerConfigure"></param>
         /// <returns></returns>
-        public static AuthenticationBuilder AddJwt(this IServiceCollection services, Action<AuthenticationOptions> configureOptions = null, object tokenValidationParameters = default)
+        public static AuthenticationBuilder AddJwt(this IServiceCollection services, Action<AuthenticationOptions> authenticationConfigure = null, object tokenValidationParameters = default, Action<JwtBearerOptions> jwtBearerConfigure = null)
         {
             // 配置 JWT 选项
             ConfigureJWTOptions(services);
@@ -66,16 +76,47 @@ namespace Microsoft.Extensions.DependencyInjection
             // 添加默认授权
             return services.AddAuthentication(options =>
              {
-                 if (configureOptions == null)
-                 {
-                     options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                     options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                 }
-                 else configureOptions.Invoke(options);
+                 options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                 options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+
+                 // 添加自定义配置
+                 authenticationConfigure?.Invoke(options);
              }).AddJwtBearer(options =>
              {
                  options.TokenValidationParameters = (tokenValidationParameters as TokenValidationParameters) ?? JWTEncryption.CreateTokenValidationParameters(jwtSettings);
+
+                 // 添加自定义配置
+                 jwtBearerConfigure?.Invoke(options);
              });
+        }
+
+        /// <summary>
+        /// 添加 JWT 授权
+        /// </summary>
+        /// <typeparam name="TAuthorizationHandler"></typeparam>
+        /// <param name="services"></param>
+        /// <param name="authenticationConfigure"></param>
+        /// <param name="tokenValidationParameters"></param>
+        /// <param name="jwtBearerConfigure"></param>
+        /// <param name="enableGlobalAuthorize"></param>
+        /// <returns></returns>
+        public static AuthenticationBuilder AddJwt<TAuthorizationHandler>(this IServiceCollection services, Action<AuthenticationOptions> authenticationConfigure = null, object tokenValidationParameters = default, Action<JwtBearerOptions> jwtBearerConfigure = null, bool enableGlobalAuthorize = false)
+            where TAuthorizationHandler : class, IAuthorizationHandler
+        {
+            // 加载 Furion 程序集
+            var furionAssembly = Assembly.Load("Furion");
+
+            // 获取添加授权类型
+            var authorizationServiceCollectionExtensionsType = furionAssembly.GetType("Microsoft.Extensions.DependencyInjection.AuthorizationServiceCollectionExtensions");
+            var addAppAuthorizationMethod = authorizationServiceCollectionExtensionsType
+                .GetMethods(BindingFlags.Public | BindingFlags.Static)
+                .Where(u => u.Name == "AddAppAuthorization" && u.IsGenericMethod && u.GetParameters().Length > 0 && u.GetParameters()[0].ParameterType == typeof(IServiceCollection)).First();
+
+            // 添加策略授权服务
+            addAppAuthorizationMethod.MakeGenericMethod(typeof(TAuthorizationHandler)).Invoke(null, new object[] { services, null, enableGlobalAuthorize });
+
+            // 添加授权
+            return services.AddJwt(authenticationConfigure, tokenValidationParameters, jwtBearerConfigure);
         }
 
         /// <summary>